fix(evdev): fix critical crash and system lockup on connection
- Fix nil pointer dereference in translateKeyboardEvent when client.Capabilities is nil
- Add panic recovery with device release in processEvents and callbacks
- Add adaptive watchdog (2s when active, 5s when idle)
- Add SIGUSR1 handler for emergency device release (kill -USR1 <pid>)
- Add ForceRelease method for external emergency recovery

+++ internal/usecase/client/client.go
@@ -346,7 +346,7 @@ func (c *UseCaseImpl) injectEvent(ctx context.Context, event *domain.InputEvent)
 		// Use semantic character injection if a character is provided
 		if event.Keyboard.Character != nil {
 			log.Debug().
-				Int32("char", int32(*event.Keyboard.Character)).
+				Int32("char", *event.Keyboard.Character).
 				Bool("pressed", event.Keyboard.Pressed).
 				Msg("Injecting keyboard character")
 			return c.inputInjection.InjectCharacter(ctx, *event.Keyboard.Character, event.Keyboard.Pressed)
